add dice in extra file dice.py
for rolling two dices use the function RollTwoDice()
returns the sum of both dices and a bool whether both dices show the same number.
the function RollTwoDiceWhichGreater() returns the sum of two dices too,
but returns instead of the bool an int which dice has the greater number.
(-1 for both same number)